+++ element_behavior/dlc_draft.py
@@ -128,6 +128,8 @@ class PoseEstimation(dj.Computed):
         """
 
     def make(self, key):
+        from .readers import dlc
+
         task_mode, output_dir = (PoseEstimationTask & key).fetch1('task_mode', 'dlc_output_dir')
         dlc_model = (Model & key).fetch1()
 
@@ -141,16 +143,16 @@ class PoseEstimation(dj.Computed):
         if task_mode == 'trigger':
             do_pose_estimation(video_filepaths, dlc_model, project_path, output_dir)
 
-        dlc_result = dlc_reader.DLCLoader(output_dir)
+        dlc_result = dlc.PoseEstimation(output_dir)
 
         body_parts = [{**key,
                        'body_part': k,
-                       'frame_index': v['frame_index'],
+                       'frame_index': np.arange(dlc_result.nframes),
                        'x_pos': v['x_pos'],
                        'y_pos': v['y_pos'],
+                       'z_pos': v.get('z_pos'),
                        'likelihood': v['likelihood']}
                       for k, v in dlc_result.data.items()]
 
         self.insert1({**key, 'post_estimation_time': dlc_result.creation_time})
         self.BodyPartPosition.insert(body_parts)
-
